refactor(definitions): move funcs from root definitions
to definitions sub-modules
