+++ ohsome_quality_analyst/api/api.py
@@ -41,13 +41,12 @@ from ohsome_quality_analyst.api.response_models import (
 from ohsome_quality_analyst.config import configure_logging
 from ohsome_quality_analyst.definitions import (
     ATTRIBUTION_URL,
-    get_indicator_definitions,
     get_metadata,
-    get_report_definitions,
-    get_topic_definition,
-    get_topic_definitions,
 )
-from ohsome_quality_analyst.indicators.definitions import IndicatorEnum
+from ohsome_quality_analyst.indicators.definitions import (
+    IndicatorEnum,
+    get_indicator_definitions,
+)
 from ohsome_quality_analyst.projects.definitions import (
     ProjectEnum,
     get_project,
@@ -58,8 +57,15 @@ from ohsome_quality_analyst.quality_dimensions.definitions import (
     get_quality_dimension,
     get_quality_dimensions,
 )
-from ohsome_quality_analyst.reports.definitions import ReportEnum
-from ohsome_quality_analyst.topics.definitions import TopicEnum
+from ohsome_quality_analyst.reports.definitions import (
+    ReportEnum,
+    get_report_definitions,
+)
+from ohsome_quality_analyst.topics.definitions import (
+    TopicEnum,
+    get_topic_definition,
+    get_topic_definitions,
+)
 from ohsome_quality_analyst.utils.exceptions import (
     HexCellsNotFoundError,
     OhsomeApiError,
