+++ ohsome_quality_analyst/indicators/definitions.py
@@ -1,5 +1,29 @@
 from enum import Enum
 
-from ohsome_quality_analyst.definitions import get_indicator_names
+from ohsome_quality_analyst.definitions import load_metadata
+from ohsome_quality_analyst.indicators.models import IndicatorMetadata
+from ohsome_quality_analyst.projects.definitions import ProjectEnum
+from ohsome_quality_analyst.topics.definitions import load_topic_definitions
+
+
+def get_indicator_names() -> list[str]:
+    return list(load_metadata("indicators").keys())
+
 
 IndicatorEnum = Enum("IndicatorEnum", {name: name for name in get_indicator_names()})
+
+
+def get_indicator_definitions(
+    project: ProjectEnum = None,
+) -> dict[str, IndicatorMetadata]:
+    indicators = load_metadata("indicators")
+    if project is not None:
+        return {k: v for k, v in indicators.items() if project in v.projects}
+    else:
+        return indicators
+
+
+def get_valid_indicators(topic_key: str) -> tuple:
+    """Get valid Indicator/Topic combination of a Topic."""
+    td = load_topic_definitions()
+    return tuple(td[topic_key].indicators)
