+++ scripts/run_mapping_saturation_models.py
@@ -11,8 +11,8 @@ from rpy2.rinterface_lib.embedded import RRuntimeError
 
 import ohsome_quality_analyst.geodatabase.client as db_client
 import ohsome_quality_analyst.ohsome.client as ohsome_client
-from ohsome_quality_analyst.definitions import get_topic_definition
 from ohsome_quality_analyst.indicators.mapping_saturation import models
+from ohsome_quality_analyst.topics.definitions import get_topic_definition
 
 
 def plot(xdata, ydata, model_list):
