+++ tests/conftest.py
@@ -7,9 +7,7 @@ from geojson import Feature, FeatureCollection, Polygon
 
 from ohsome_quality_analyst.definitions import (
     get_metadata,
-    get_topic_definition,
     load_metadata,
-    load_topic_definitions,
 )
 from ohsome_quality_analyst.indicators.models import IndicatorMetadata
 from ohsome_quality_analyst.projects.definitions import get_project, load_projects
@@ -20,6 +18,10 @@ from ohsome_quality_analyst.quality_dimensions.definitions import (
 )
 from ohsome_quality_analyst.quality_dimensions.models import QualityDimension
 from ohsome_quality_analyst.reports.models import ReportMetadata
+from ohsome_quality_analyst.topics.definitions import (
+    get_topic_definition,
+    load_topic_definitions,
+)
 
 # from ohsome_quality_analyst.indicators import MappingSaturation
 # from ohsome_quality_analyst.indicators.models import BaseIndicator as Indicator
