+++ tests/integrationtests/utils.py
@@ -5,7 +5,7 @@ from unittest.mock import MagicMock
 import geojson
 import vcr
 
-from ohsome_quality_analyst.definitions import get_topic_definition
+from ohsome_quality_analyst.topics.definitions import get_topic_definition
 from ohsome_quality_analyst.topics.models import TopicDefinition
 
 TEST_DIR = os.path.dirname(os.path.abspath(__file__))
